refactor(llm): move Ollama model support logic to dedicated helper and strengthen tests
